Move file from pages into their own folders
