+++ launchpad-frontend/src/index.js
@@ -4,8 +4,8 @@ import "./index.css";
 import { NavBar } from "./navigation";
 import reportWebVitals from "./reportWebVitals";
 import { LandingPage } from "./landingPage/landingPage";
-import { AccountSettings } from "./pages/accountSettings";
-import { MyApplication } from "./pages/myApplication";
+import { AccountSettings } from "./accountSettingPage/accountSettings";
+import { MyApplication } from "./myApplicationPage/myApplication";
 import { NotificationsPage } from "./notificationsPage/notificationsPage";
 import { JobPostings } from "./jobsPage/jobsPage";
 import { App } from "./App";
