+++ launchpad-frontend/src/myApplicationPage/myApplication.js
